restructure multileadtype and remove big test file
